backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
boto3
pika
python-multipart
passlib[bcrypt]
python-jose[cryptography]
email-validator
bcrypt==4.0.1
requests
elasticsearch
redis
numpy
minio
opencv-python-headless
docker